detatch android related downloader from crawler
