+++ app/src/main/java/org/schabi/newpipe/services/StreamingService.java
@@ -25,7 +25,7 @@ public interface StreamingService {
         public String name = "";
     }
     ServiceInfo getServiceInfo();
-    VideoExtractor getExtractorInstance(String url);
+    VideoExtractor getExtractorInstance(String url, Downloader downloader);
     SearchEngine getSearchEngineInstance();
 
     /**When a VIEW_ACTION is caught this function will test if the url delivered within the calling
